Code/requirements.txt
Werkzeug==2.3.6
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.3
greenlet==2.0.2
typing-extensions==4.6.3
MarkupSafe==2.1.3
six==1.16.0
pytz==2023.3
aniso8601==9.0.1
numpy==1.25.0
tzdata==2023.3
pandas==2.0.2
python-dateutil==2.8.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
passlib==1.7.4
Flask-Principal==0.4.0
Flask-BabelEx==0.9.4
speaklater==1.3
Babel==2.12.1
blinker==1.6.2
WTForms==3.0.1
email_validator
flask-security-too
bcrypt
requests
mailhog
weasyprint==52.5
pdfkit
schedule
sendgrid
Flask-Caching
Flask==2.2.3
Flask-SQLALchemy==3.0.3
flask-restful==0.3.10
Flask-JWT-Extended
Flask-cors
python-dotenv
celery
redis